requirements.txt
aiogram==3.4.1
flask>=2.0
psutil
python-dotenv
aiofiles
